Modified background color back from purple to white in styles.css
Added a property called `isListenerAttached` in the constructor as a boolean flag init to false. When `removedTaskUI()` is called checks if the flag is false and if so attaches the event listener and sets the flag to true. Used to prevent subsequent calls of the listener event called again in project.js.